refactor cms pricing form
